igore sign-up and sign-in routes if user is authenticated
If the user is already signed-in, there's no need for them to reauthenticate. So we add an `onEnter` hook to the route to check to see if there is an existing user and token. If there is, we will redirect to wherever the user should be. In most cases, that will be the root route. But in some cases, if `redirect` and / or `responseType` are set in the query string, we'll redirect to wherever they point, possibly including the token in the query string if `redirectType === 'token'`.